+++ poem/src/endpoint/mod.rs
@@ -8,14 +8,14 @@ mod catch_all_error;
 mod catch_error;
 #[allow(clippy::module_inception)]
 mod endpoint;
-#[cfg(feature = "static-files")]
-mod files;
 mod inspect_all_err;
 mod inspect_err;
 mod map;
 mod map_to_response;
 #[cfg(feature = "prometheus")]
 mod prometheus_exporter;
+#[cfg(feature = "static-files")]
+mod static_files;
 #[cfg(feature = "tower-compat")]
 mod tower_compat;
 
@@ -26,13 +26,13 @@ pub use before::Before;
 pub use catch_all_error::CatchAllError;
 pub use catch_error::CatchError;
 pub use endpoint::{make, make_sync, BoxEndpoint, Endpoint, EndpointExt, IntoEndpoint};
-#[cfg(feature = "static-files")]
-pub use files::{StaticFile, StaticFiles};
 pub use inspect_all_err::InspectAllError;
 pub use inspect_err::InspectError;
 pub use map::Map;
 pub use map_to_response::MapToResponse;
 #[cfg(feature = "prometheus")]
 pub use prometheus_exporter::PrometheusExporter;
+#[cfg(feature = "static-files")]
+pub use static_files::{StaticFileEndpoint, StaticFilesEndpoint};
 #[cfg(feature = "tower-compat")]
 pub use tower_compat::TowerCompatExt;