+++ poem-openapi/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-- [1.2.25] 2022-1-4
+# [1.2.26] 2022-1-7
+
+- Add `StaticFile` response.
+
+# [1.2.25] 2022-1-4
 
 - Add `deprecated` attribute to `Enum` macro.