+++ poem/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.2.27] 2022-1-7
+
+- Rename `poem::endpoint::StaticFiles` to `poem::endpoint::StaticFilesEndpoint`.
+- Rename `poem::endpoint::StaticFile` to `poem::endpoint::StaticFileEndpoint`.
+- Add `poem::web::StaticFileRequest` extractor.
+
 # [1.2.26] 2022-1-6
 
 - Add I18N support with [`fluent`](https://crates.io/crates/fluent).